requirements.txt
dragonfly2
dragonfly2[kaldi]
